feat: Migrate OptionsSheet and Update Dependencies
This commit migrates the `OptionsSheet` components to the `kmpuiviews` module for multiplatform support and updates various parts of the application to use the new components. Additionally, it addresses some related issues and updates dependencies.

-   **OptionsSheet Migration:**
    - Moved `OptionsSheetValues`, `OptionsSheetScope`, and `KmpItemModelOptionsSheet` to `kmpuiviews` to be multiplatform classes.
    - Updated all usages of the old `OptionsSheet` to use the new one in `kmpuiviews`.
    - Removed the old `OptionsSheet` classes from `UIViews`.
    - Removed unused `OptionsSheetValues` implementation from `OtakuCustomListScreen` and `NotificationFragment`
-   - Updated `OtakuCustomListScreen` and `NotificationFragment` to use `kmpuiviews` `OptionsSheetValues`.
- **Dependencies:**
+   - Updated `OtakuCustomListScreen` and `NotificationFragment` to use the new `OptionsSheetValues` from `kmpuiviews`.
+
+- **Dependencies:**
     - Updated the Haze dependency to version 1.6.0-beta02.
-    - Updated imports to be multiplatform
+- **Other:**
+    - Updated `search` in `ApiService` to use KMP models.
+    - Updated `ItemModel` mappings in `AndroidApiService`.
+    - Updated imports to be multiplatform.
     - Removed unused imports.
+    - Removed unused `optionsSheetList` from `UIViews`.
