+++ gradle/libs.versions.toml
@@ -16,7 +16,7 @@ dragselect = "2.4.3"
 easylauncher = "6.4.0"
 firebaseCrashlyticsGradle = "3.0.3"
 googleServices = "4.4.2"
-haze = "1.6.0-beta01"
+haze = "1.6.0-beta02"
 kamelImage = "1.0.3"
 latestAboutLibsRelease = "12.0.1"
 coroutinesVersion = "1.10.2"
