+++ kmpmodels/src/androidMain/kotlin/com/programmersbox/kmpmodels/Platform.android.kt
@@ -353,11 +353,7 @@ class ModelMapper(private val application: Application) {
             override suspend fun sourceByUrl(url: String): ItemModel = mapItemModel(apiService.sourceByUrl(url))
 
             override suspend fun search(searchText: CharSequence, page: Int, list: List<ItemModel>): List<ItemModel> =
-                apiService.search(
-                    searchText,
-                    page,
-                    list.map { mapItemModel(it) }
-                ).map { mapItemModel(it) }
+                apiService.search(searchText, page, list.map { mapItemModel(it) }).map { mapItemModel(it) }
 
             override fun searchListFlow(searchText: CharSequence, page: Int, list: List<ItemModel>): Flow<List<ItemModel>> = apiService
                 .searchListFlow(searchText, page, list.map { mapItemModel(it) })
